+++ packages/cli/src/lib/autorun/autorun-command.ts
@@ -6,8 +6,8 @@ import {
   collectAndPersistReports,
   upload,
 } from '@code-pushup/core';
-import { CLI_NAME } from '../cli';
-import { yargsOnlyPluginsOptionsDefinition } from '../implementation/only-plugins.options';
+import { CLI_NAME } from '../constants';
+import { yargsOnlyPluginsOptionsDefinition } from '../implementation/only-plugins-options';
 
 type AutorunOptions = CollectOptions & UploadOptions;
 
@@ -34,10 +34,11 @@ export function yargsAutorunCommandObject() {
       };
 
       await collectAndPersistReports(optionsWithFormat);
-      if (!options.upload) {
-        console.warn('Upload skipped because configuration is not set.'); // @TODO log verbose
-      } else {
+
+      if (options.upload) {
         await upload(options);
+      } else {
+        console.warn('Upload skipped because configuration is not set.'); // @TODO log verbose
       }
     },
   } satisfies CommandModule;