+++ packages/utils/src/index.ts
@@ -26,17 +26,22 @@ export {
   pluralizeToken,
   slugify,
   truncateDescription,
+  truncateIssueMessage,
   truncateText,
   truncateTitle,
 } from './lib/formatting';
 export { getLatestCommit, git } from './lib/git';
+export { groupByStatus } from './lib/group-by-status';
 export {
   isPromiseFulfilledResult,
   isPromiseRejectedResult,
 } from './lib/guards';
 export { logMultipleResults } from './lib/log-results';
-export { NEW_LINE } from './lib/md';
 export { ProgressBar, getProgressBar } from './lib/progress';
+export { generateMdReport } from './lib/reports/generate-md-report';
+export { generateStdoutSummary } from './lib/reports/generate-stdout-summary';
+export { ScoredReport, scoreReport } from './lib/reports/scoring';
+export { sortReport } from './lib/reports/sorting';
 export {
   CODE_PUSHUP_DOMAIN,
   FOOTER_PREFIX,
@@ -44,10 +49,7 @@ export {
   calcDuration,
   compareIssueSeverity,
   loadReport,
-} from './lib/report';
-export { reportToMd } from './lib/report-to-md';
-export { reportToStdout } from './lib/report-to-stdout';
-export { ScoredReport, scoreReport } from './lib/scoring';
+} from './lib/reports/utils';
 export {
   CliArgsObject,
   countOccurrences,
@@ -60,4 +62,3 @@ export {
   toUnixPath,
 } from './lib/transform';
 export { verboseUtils } from './lib/verbose-utils';
-export { groupByStatus } from './lib/group-by-status';