+++ packages/utils/src/lib/file-system.ts
@@ -68,7 +68,10 @@ export class NoExportError extends Error {
   }
 }
 
-export async function importEsmModule(options: Options): Promise<unknown> {
+export async function importEsmModule<T = unknown>(
+  options: Options,
+  parseFn?: (mod: unknown) => T,
+): Promise<T> {
   const { mod } = await bundleRequire({
     format: 'esm',
     ...options,
@@ -78,7 +81,7 @@ export async function importEsmModule(options: Options): Promise<unknown> {
     throw new NoExportError(options.filepath);
   }
 
-  return mod.default;
+  return parseFn ? parseFn(mod.default) : mod.default;
 }
 
 export function pluginWorkDir(slug: string): string {