+++ packages/utils/src/lib/reports/constants.ts
@@ -1,0 +1,6 @@
+export const NEW_LINE = '\n';
+
+export const SCORE_COLOR_RANGE = {
+  GREEN_MIN: 0.9,
+  YELLOW_MIN: 0.5,
+};